+++ sdk/formrecognizer/azure-ai-formrecognizer/samples/v3.2/async_samples/sample_classify_document_async.py
@@ -77,7 +77,7 @@ async def main():
         from azure.ai.formrecognizer.aio import DocumentModelAdministrationClient
         from azure.ai.formrecognizer import (
             ClassifierDocumentTypeDetails,
-            AzureBlobSource,
+            BlobSource,
         )
 
         endpoint = os.environ["AZURE_FORM_RECOGNIZER_ENDPOINT"]
@@ -91,13 +91,13 @@ async def main():
             poller = await document_model_admin_client.begin_build_document_classifier(
                 doc_types={
                     "IRS-1040-A": ClassifierDocumentTypeDetails(
-                        source=AzureBlobSource(
+                        source=BlobSource(
                             container_url=blob_container_sas_url,
                             prefix="IRS-1040-A/train",
                         )
                     ),
                     "IRS-1040-D": ClassifierDocumentTypeDetails(
-                        source=AzureBlobSource(
+                        source=BlobSource(
                             container_url=blob_container_sas_url,
                             prefix="IRS-1040-D/train",
                         )