+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_inputs_outputs/base.py
@@ -11,7 +11,7 @@ class _InputOutputBase(DictMixin, RestTranslatableMixin):
         *,
         type,  # pylint: disable=redefined-builtin
         **kwargs,  # pylint: disable=unused-argument
-    ):
+    ) -> None:
         """Base class for Input & Output class.
 
         This class is introduced to support literal output in the future.