+++ sdk/ml/azure-ai-ml/azure/ai/ml/operations/_workspace_operations.py
@@ -199,8 +199,8 @@ class WorkspaceOperations(WorkspaceOperationsBase):
             i.e., container registry, storage account, key vault, and application insights.
             The default is False. Set to True to delete these resources.
         :type delete_dependent_resources: bool
-        :keyword permanently_delete: Workspaces are soft-deleted state by default to allow recovery of workspace data.
-            Set this flag to override the soft-delete behavior and permanently delete your workspace.
+        :param permanently_delete: Workspaces are soft-deleted by default to allow recovery of workspace data.
+            Set this flag to true to override the soft-delete behavior and permanently delete your workspace.
         :type permanently_delete: bool
         :return: A poller to track the operation status.
         :rtype: ~azure.core.polling.LROPoller[None]