+++ sdk/search/azure-search-documents/tests/test_search_client_buffered_sender_live.py
@@ -20,8 +20,8 @@ class TestSearchIndexingBufferedSender(AzureRecordedTestCase):
     @search_decorator(schema="hotel_schema.json", index_batch="hotel_small.json")
     @recorded_by_proxy
     def test_search_client_index_buffered_sender(self, endpoint, api_key, index_name):
-        client = SearchClient(endpoint, index_name, api_key)
-        batch_client = SearchIndexingBufferedSender(endpoint, index_name, api_key)
+        client = SearchClient(endpoint, index_name, api_key, retry_backoff_factor=60)
+        batch_client = SearchIndexingBufferedSender(endpoint, index_name, api_key, retry_backoff_factor=60)
         try:
             doc_count = 10
             doc_count = self._test_upload_documents_new(client, batch_client, doc_count)