+++ sdk/search/azure-search-documents/tests/test_search_index_client_data_source_live.py
@@ -31,7 +31,7 @@ class TestSearchClientDataSources(AzureRecordedTestCase):
     @recorded_by_proxy
     def test_data_source(self, endpoint, api_key, **kwargs):
         storage_cs = kwargs.get("search_storage_connection_string")
-        client = SearchIndexerClient(endpoint, api_key)
+        client = SearchIndexerClient(endpoint, api_key, retry_backoff_factor=60)
         self._test_create_datasource(client, storage_cs)
         self._test_delete_datasource(client, storage_cs)
         self._test_get_datasource(client, storage_cs)